+++ src/api/api.ts
@@ -204,11 +204,14 @@ export default class MastoApi {
 
     // Get accounts the user is following
     async getFollowedAccounts(): Promise<Account[]> {
-        return await this.getApiRecords<mastodon.v1.Account>({
+        const accounts = await this.getApiRecords<mastodon.v1.Account>({
             fetch: this.api.v1.accounts.$select(this.user.id).following.list,
             storageKey: StorageKey.FOLLOWED_ACCOUNTS,
             maxRecords: Config.maxFollowingAccountsToPull,
         }) as Account[];
+
+        accounts.forEach(account => account.isFollowed = true);
+        return accounts;
     }
 
     // Get hashtags the user is following