+++ src/api/objects/account.ts
@@ -84,7 +84,7 @@ export default class Account implements AccountObj {
         accountObj.suspended = account.suspended || false;
         accountObj.roles = account.roles || [];
         // Fedialgo extension fields
-        accountObj.isFollowed = false;
+        accountObj.isFollowed = false;  // Must be set later, in Toot.complete() etc.
         accountObj.webfingerURI = accountObj.buildWebfingerURI();
         return accountObj;
     }